+++ lotti/lib/pages/dashboards/dashboards_list_page.dart
@@ -67,7 +67,7 @@ class DashboardCard extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Card(
-      color: AppColors.headerBgColor,
+      color: AppColors.entryCardColor,
       elevation: 8.0,
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(8.0),