+++ lotti/lib/pages/settings/tags/tags_page.dart
@@ -137,7 +137,7 @@ class TagCard extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Card(
-      color: AppColors.headerBgColor,
+      color: AppColors.entryCardColor,
       elevation: 8.0,
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(8.0),