+++ lotti/lib/widgets/journal/entry_details_widget.dart
@@ -126,7 +126,7 @@ class _EntryDetailWidgetState extends State<EntryDetailWidget> {
           child: ClipRRect(
             borderRadius: BorderRadius.circular(8.0),
             child: Container(
-              color: AppColors.headerBgColor,
+              color: AppColors.entryCardColor,
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [