+++ lotti/lib/widgets/sync/qr_widget.dart
@@ -27,6 +27,7 @@ class EncryptionQrWidget extends StatelessWidget {
       return Padding(
         padding: const EdgeInsets.all(8.0),
         child: Column(
+          mainAxisSize: MainAxisSize.min,
           children: [
             Button(
               localizations.settingsSyncGenKeyButton,